функции работы с BD вынесены в файл data_functions.php
