+++ mysql_helper.php
@@ -9,7 +9,11 @@ mysqli_set_charset($link, "utf8");
 if (!$link) {
     $error = mysqli_connect_error();
     $layout_content = include_template('error.php', ['error' => $error]);
+
+    print($layout_content);
+    exit(1);
 }
+
 /**
  * Создает подготовленное выражение на основе готового SQL запроса и переданных данных
  *
